+++ src/main/kotlin/de/visualdigits/kaudiotagger/model/id3/types/PictureTypes.kt
@@ -38,9 +38,9 @@ enum class PictureTypes(
 
         fun fromId(id: Int?): PictureTypes? = entries.find { e -> e.id == id }
 
-        fun getValueToIdMap(): Map<String, Int> = entries.associate { e -> Pair(e.friendlyName, e.id) }
+        fun getValueToIdMap(): Map<String, Long> = entries.associate { e -> Pair(e.friendlyName, e.id.toLong()) }
 
-        fun getIdToValueMap(): Map<Int, String> = entries.associate { e -> Pair(e.id, e.friendlyName) }
+        fun getIdToValueMap(): Map<Long, String> = entries.associate { e -> Pair(e.id.toLong(), e.friendlyName) }
 
     }
 }
